code/requirements.txt
arxiv
httpx
slack_sdk
google-generativeai
functions_framework